% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-cols.R
\name{add_cols}
\alias{add_cols}
\title{Add new columns to a latin rectangle}
\usage{
add_cols(R, cols, l_order, strategy = next_col_matching)
}
\arguments{
\item{R}{A latin rectangle}

\item{cols}{Indices of columns to add}

\item{l_order}{Dimension of latin square}

\item{strategy}{Strategy for filling columns}
}
\value{
A latin rectangle
}
\description{
Add new columns to a latin rectangle
}
