% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-rows.R
\name{add_rows}
\alias{add_rows}
\title{Embed latin rectangle in a latin square}
\usage{
add_rows(R, rows)
}
\arguments{
\item{R}{latin rectangle}

\item{rows}{empty rows to be filled}
}
\value{
A latin rectangle with more rows.
}
\description{
Input is a latin rectangle as a data frame with
variables for row, column and symbol. Output is
a latin square in the same format which contains
the given latin rectangle in the first rows.
}
\details{
Use can optionally provide a vector of row indices.
Only those rows will be filled if that optional
vector is provided.
}
