% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-latin-square.R
\name{create_latin_square}
\alias{create_latin_square}
\title{Generate latin squares}
\usage{
create_latin_square(l_order)
}
\arguments{
\item{l_order}{Latin square order.}
}
\value{
A latin square.
}
\description{
Create a latin square by first generating a
1 x l_order latin rectangle and then embedding
it in a latin square.
}
