% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-tbl.R
\name{edge_tbl}
\alias{edge_tbl}
\title{Symbols missing from columns edge data frame}
\usage{
edge_tbl(R, i, l_order = 3)
}
\arguments{
\item{R}{latin rectangle}

\item{i}{column}

\item{l_order}{size of latin square R is going to be embedded into}
}
\value{
The edge data frame.
}
\description{
Constructs a data frame representing the edges
of a bipartite graph based on a latin rectangle
where the graph has an edge for every symbol not
already used in a column.
}
\details{
Acutally, this is just for one column.
}
