% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-tbl.R
\name{edge_tbl_2}
\alias{edge_tbl_2}
\title{Symbols missing from rows data frame}
\usage{
edge_tbl_2(R, i, l_order = 3)
}
\arguments{
\item{R}{latin rectangle}

\item{i}{row index}

\item{l_order}{size of latin square R is going to be embedded into}
}
\value{
The edge data frame.
}
\description{
Symbols missing from rows data frame
}
