% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-row.R
\name{first_row_natural}
\alias{first_row_natural}
\title{First row in in natural order}
\usage{
first_row_natural(n)
}
\arguments{
\item{n}{Number of columns}
}
\value{
A 1 x n latin rectangle with first row 1,...,n
}
\description{
First row in in natural order
}
