% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next-col.R
\name{next_col_matching}
\alias{next_col_matching}
\title{Matching strategy for adding new columns}
\usage{
next_col_matching(R, i, l_order)
}
\arguments{
\item{R}{a latin rectangle}

\item{i}{column index}

\item{l_order}{dimension}
}
\value{
a latin rectangle with more columns
}
\description{
Matching strategy for adding new columns
}
