% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next-col.R
\name{next_col_random}
\alias{next_col_random}
\title{Random strategy for choosing a new columns}
\usage{
next_col_random(R, i, l_order)
}
\arguments{
\item{R}{a latin rectangle}

\item{i}{column index}

\item{l_order}{dimension}
}
\value{
A latin rectangle with more columns
}
\description{
Random strategy for choosing a new columns
}
