% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next-row.R
\name{next_row_matching}
\alias{next_row_matching}
\title{Find a compatible row for extending a latin rectangle}
\usage{
next_row_matching(R, i, l_order)
}
\arguments{
\item{R}{A latin rectangle}

\item{i}{Number of columns to add}

\item{l_order}{Order of R}
}
\value{
A latin rectangle with more rows.
}
\description{
Given an input latin rectangle this function will
generate a new row that can be added to the latin
rectangle.
}
\details{
The method used is to create a bipartite graph
with vertex partitions for columns and symbols
missing from columns and then find a maximum
matching in that bipartite graph.
}
