% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next-row.R
\name{next_row_random}
\alias{next_row_random}
\title{Find a random new row for a latin rectangle}
\usage{
next_row_random(R, i, l_order)
}
\arguments{
\item{R}{A latin rectangle}

\item{i}{Number of columns to add}

\item{l_order}{Order of R}
}
\value{
A latin square with more rows.
}
\description{
Find a random new row for a latin rectangle
}
