% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-tidygraph.R
\name{to_tidygraph}
\alias{to_tidygraph}
\title{Symbols missing from columns bipartite graph}
\usage{
to_tidygraph(R, l_order = 3)
}
\arguments{
\item{R}{A latin rectangle.}

\item{l_order}{Order of R.}
}
\value{
A bipartite graph.
}
\description{
Input is a latin rectangle as a data frame with
variables for row, column and symbol. Output is
a tidygraph representing the bipartite graph with
vertices for columns and symbols and edges representing
symbols missing from columns.
}
