% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-tidygraph.R
\name{to_tidygraph_2}
\alias{to_tidygraph_2}
\title{Symbols missing from rows bipartite graph}
\usage{
to_tidygraph_2(R, l_order, n_rows, n_cols)
}
\arguments{
\item{R}{A latin rectangle.}

\item{l_order}{Order of R.}

\item{n_rows}{Number of rows.}

\item{n_cols}{Number of columns.}
}
\value{
A bipartite graph.
}
\description{
Symbols missing from rows bipartite graph
}
